renames repo delete command to tkn pac delete repo
