+++ pkg/cmd/tknpac/delete/root.go
@@ -1,0 +1,23 @@
+package delete
+
+import (
+	"github.com/openshift-pipelines/pipelines-as-code/pkg/cli"
+	"github.com/openshift-pipelines/pipelines-as-code/pkg/params"
+	"github.com/spf13/cobra"
+)
+
+func Root(clients *params.Run, ioStreams *cli.IOStreams) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:          "delete",
+		Aliases:      []string{"rm"},
+		Short:        "Delete Pipelines as Code resources",
+		Long:         `Delete Pipelines as Code resources`,
+		SilenceUsage: true,
+		Annotations: map[string]string{
+			"commandType": "main",
+		},
+	}
+
+	cmd.AddCommand(repositoryCommand(clients, ioStreams))
+	return cmd
+}
