+++ pkg/cmd/tknpac/root.go
@@ -5,6 +5,7 @@ import (
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/cmd/tknpac/bootstrap"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/cmd/tknpac/completion"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/cmd/tknpac/create"
+	"github.com/openshift-pipelines/pipelines-as-code/pkg/cmd/tknpac/delete"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/cmd/tknpac/describe"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/cmd/tknpac/generate"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/cmd/tknpac/list"
@@ -32,6 +33,7 @@ func Root(clients *params.Run) *cobra.Command {
 	cmd.AddCommand(version.Command(ioStreams))
 	cmd.AddCommand(create.Root(clients, ioStreams))
 	cmd.AddCommand(list.Root(clients, ioStreams))
+	cmd.AddCommand(delete.Root(clients, ioStreams))
 	cmd.AddCommand(describe.Root(clients, ioStreams))
 	cmd.AddCommand(repository.Root(clients, ioStreams))
 	cmd.AddCommand(resolve.Command(clients, ioStreams))
